Wrote the new testing function
added a boolean Crit failed to the struct that keeps the number of
passed and tested files.  It will need to be initialized to false, and
if it ever becomes true it should stay true.  The function needs the
complete path to the executable, the name of the test case, the
directory of the test case, the student directory, the student record,
and the student log.